+++ src/views/UserDetail/UserDetail.js
@@ -499,7 +499,7 @@ class UserDetail extends React.Component {
     const showActionMenu = stripes.hasPerm('ui-users.edit')
       || stripes.hasPerm('ui-users.patron-blocks.all')
       || stripes.hasPerm('ui-users.feesfines.actions.all')
-      || stripes.hasPerm('ui-requests.all')
+      || stripes.hasPerm('ui-requests.create')
       || stripes.hasPerm('ui-users.delete,ui-users.open-transactions.view')
       || stripes.hasPerm('ui-users.profile-pictures.all');